Discord Update and bug fixes
Discord.js released a major update which meant the code needed updating.

I have amended all the code to reflect the changes, fixed some bugs and restructured some of the code for ease and fixing.

Also created a guildObj and memberObj for the events to cut down on repetitive typing, and put all embeds into a class to reduce clutter across files.